新增RxHttp代码类
