+++ rxhttp/src/main/java/rxhttp/RxHttpRetry.kt
@@ -1,10 +1,7 @@
 package rxhttp
 
-import io.reactivex.Scheduler
-import io.reactivex.functions.Consumer
 import kotlinx.coroutines.delay
 import okhttp3.OkHttpClient
-import rxhttp.wrapper.entity.Progress
 import rxhttp.wrapper.parse.Parser
 
 /**
@@ -14,24 +11,11 @@ import rxhttp.wrapper.parse.Parser
  * Time: 17:06
  */
 class RxHttpRetry(
-    private val baseRxHttp: BaseRxHttp,
+    baseRxHttp: BaseRxHttp,
     private var times: Int = 0,
     private val period: Long = 0L,
     private val test: ((Throwable) -> Boolean)? = null
-) : BaseRxHttp() {
-
-    override val breakDownloadOffSize: Long
-        get() = baseRxHttp.breakDownloadOffSize
-
-    override fun buildRequest() = baseRxHttp.buildRequest()
-
-    override fun <T> asParser(parser: Parser<T>) = baseRxHttp.asParser(parser)
-
-    override fun asDownload(
-        destPath: String,
-        progressConsumer: Consumer<Progress<String>>,
-        observeOnScheduler: Scheduler?
-    ) = baseRxHttp.asDownload(destPath, progressConsumer, observeOnScheduler)
+) : RxHttpProxy(baseRxHttp) {
 
     override suspend fun <T> await(client: OkHttpClient, parser: Parser<T>): T {
         return try {
@@ -49,15 +33,4 @@ class RxHttpRetry(
         }
     }
 }
-
-/**
- * @param times  重试次数, 默认Int.MAX_VALUE 代表不断重试
- * @param period 重试周期, 默认为0, 单位: milliseconds
- * @param test   重试条件, 默认为空，即无条件重试
- */
-fun BaseRxHttp.retry(
-    times: Int = Int.MAX_VALUE,
-    period: Long = 0,
-    test: ((Throwable) -> Boolean)? = null
-) = RxHttpRetry(this, times, period, test)
